Numbers.java file added to Lesson 2 branch and updated to function
Create a private integer field to call randomNum with public getters and setters for that field. Added generateNumber() using Random class from the java.util package to generate a random number between 0 and 100 and set that value to the randomNum field.  Added compareNumber(int guess) that accepts an integer as a guess and returns a boolean of true or false.