GamePlay.java file added to Lesson 2 branch and updated to function
Create a private field of the type Person. Ask user for their last name and determine if last name is applied or not. Create a new instance of the Numbers.java class and generate a random number using that instance. Create a loop that prompts the player, by name, to enter a guess at the random number and compare it to the correct number selected by compareNumber(int guess).